Too much changes due to non stop programming
First of all, the backend, in this one i have modified some controllers
to accomplish the bussines rules, the docker-compose to have the ports
in local too, etc. etc. etc.

In the database some minor changes because some typos, and every one now
have a command to change to the database because in docker apparently
it's needed.

In frontend, yeah now i have a frontend, there are the endpoint for the
login, the salesmans area, the inventory area, with his request to
database and working almost everything of that endpoints.

To do next maybe, reports and employees registration, and if i can
the warehouse area because it's needed the endpoint for the CRUD of the
products, it's not sufficient with the InventoryController.